requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
httpx
fal-client
requests
python-dotenv
